Produtos e formas de pagamentos concluídos

+++ src/services/category.service.ts
@@ -1,13 +1,17 @@
 import { NotFoundError } from "../errors/not-found.error.js";
+import { ValidationError } from "../errors/validation.error.js";
 import { Category } from '../models/category.models.js';
 import { CategoryRepository } from '../repository/category.respository.js';
+import { ProductRepository } from "../repository/produt.repository.js";
 
 export class CategoryService {
 
     private categoryRepository: CategoryRepository;
+    private productRepository: ProductRepository;
 
     constructor() {
         this.categoryRepository = new CategoryRepository();
+        this.productRepository = new ProductRepository();
     }
 
     async getAll(): Promise<Category[]> {
@@ -40,6 +44,9 @@ export class CategoryService {
     }
 
     async deleteById(id: string): Promise<void> {
+        if (await this.productRepository.getCountByCategoria(id) > 0) {
+            throw new ValidationError("Não é possível excluir uma categoria com produtos relacionados!");
+        }
         return this.categoryRepository.deleteById(id);
     }
